починил выбор города

+++ routes/api.php
@@ -1,6 +1,6 @@
 <?php
 
-use App\Http\Controllers\Admin\CityController;
+use App\Http\Controllers as C;
 use App\Http\Controllers\Admin\CategoryController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
@@ -23,21 +23,21 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
 Route::group([
     'prefix' => 'admin',
     'middleware' => [
-//        'auth:sanctum',
-//        'admin'
+        'auth:sanctum',
+        'admin'
     ]
 ], function() {
     Route::group(['prefix' => 'cities'], function() {
-        Route::get('/all', [CityController::class, 'all']);
-        Route::get('/', [CityController::class, 'index'])->name('admin.city.index');
-        Route::post('/', [CityController::class,'store'])->name('admin.city.store');
-        Route::delete('/{city}', [CityController::class, 'destroy'])->name('admin.city.destroy');
-        Route::patch('/{city}', [CityController::class, 'update'])->name('admin.city.update');
+        Route::get('/all', [C\Admin\CityController::class, 'all']);
+        Route::get('/', [C\Admin\CityController::class, 'index'])->name('admin.city.index');
+        Route::post('/', [C\Admin\CityController::class,'store'])->name('admin.city.store');
+        Route::delete('/{city}', [C\Admin\CityController::class, 'destroy'])->name('admin.city.destroy');
+        Route::patch('/{city}', [C\Admin\CityController::class, 'update'])->name('admin.city.update');
 
-        Route::get('/controls', [CityController::class, 'controls'])->name('admin.city.controls.create');
-        Route::get('/controls/{city}', [CityController::class, 'controls'])->name('admin.city.controls.update');
+        Route::get('/controls', [C\Admin\CityController::class, 'controls'])->name('admin.city.controls.create');
+        Route::get('/controls/{city}', [C\Admin\CityController::class, 'controls'])->name('admin.city.controls.update');
 
-        Route::get('/{city}', [CityController::class, 'show'])->name('admin.city.show');
+        Route::get('/{city}', [C\Admin\CityController::class, 'show'])->name('admin.city.show');
 
     });
 
@@ -53,3 +53,5 @@ Route::group([
         Route::get('/{category}', [CategoryController::class, 'show'])->name('admin.categories.show');
     });
 });
+
+Route::get('/cities/all', [C\CityController::class, 'all']);
